+++ truck-polymesh/src/structured_mesh.rs
@@ -27,7 +27,7 @@ impl StructuredMesh {
 
     /// Creates a structured polygon without `uv_division` and `normal`.
     #[inline(always)]
-    pub fn from_positions_unchecked(positions: Vec<Vec<Point3>>) -> StructuredMesh {
+    pub const fn from_positions_unchecked(positions: Vec<Vec<Point3>>) -> StructuredMesh {
         StructuredMesh {
             positions,
             uv_division: None,
@@ -59,7 +59,7 @@ impl StructuredMesh {
     }
     /// Creates a structured polygon without normals.
     #[inline(always)]
-    pub fn from_positions_and_uvs_unchecked(
+    pub const fn from_positions_and_uvs_unchecked(
         positions: Vec<Vec<Point3>>,
         uv_divisions: (Vec<f64>, Vec<f64>),
     ) -> StructuredMesh {
@@ -92,7 +92,7 @@ impl StructuredMesh {
     }
     /// Creates a structured polygon without uv divisions.
     #[inline(always)]
-    pub fn from_positions_and_normals_unchecked(
+    pub const fn from_positions_and_normals_unchecked(
         positions: Vec<Vec<Point3>>,
         normals: Vec<Vec<Vector3>>,
     ) -> StructuredMesh {
@@ -135,7 +135,7 @@ impl StructuredMesh {
     /// Creates new structured mesh.
     /// Does not check whether the size of vectors are compatible before creation.
     #[inline(always)]
-    pub fn new_unchecked(
+    pub const fn new_unchecked(
         positions: Vec<Vec<Point3>>,
         uv_division: (Vec<f64>, Vec<f64>),
         normals: Vec<Vec<Vector3>>,
@@ -149,7 +149,7 @@ impl StructuredMesh {
 
     /// Returns the matrix of all positions.
     #[inline(always)]
-    pub fn positions(&self) -> &Vec<Vec<Point3>> { &self.positions }
+    pub const fn positions(&self) -> &Vec<Vec<Point3>> { &self.positions }
 
     /// Returns the vector of the mutable references to the rows of the positions matrix.
     #[inline(always)]
@@ -173,7 +173,7 @@ impl StructuredMesh {
 
     /// Returns the matrix of all normals.
     #[inline(always)]
-    pub fn normals(&self) -> Option<&Vec<Vec<Vector3>>> { self.normals.as_ref() }
+    pub const fn normals(&self) -> Option<&Vec<Vec<Vector3>>> { self.normals.as_ref() }
 
     /// Returns the vector of the mutable references to the rows of the normals matrix.
     #[inline(always)]